moved a lot of the items to WeatherAPI.js file, was able to put the city name at the top of /cityweather and the weather icon as well
